changed some things for the sake of flow
Changed a possibly awkward use of e.g. with a parenthetical, moved some words around for the first line of "Online Socialist Activism" and included the word online for emphasis on which discourse is being talked about, and changed one word from a noun used as an adjective to the adjective form of the word.